Go: split main into its own file
